+++ pref/lib/src/checkbox.dart
@@ -77,7 +77,12 @@ class _PrefCheckboxState extends State<PrefCheckbox> {
 
   @override
   Widget build(BuildContext context) {
-    var value = PrefService.of(context).getBool(widget.pref);
+    bool value;
+    try {
+      value = PrefService.of(context).get(widget.pref);
+    } catch (e) {
+      print('Unable to load the value: $e');
+    }
     if (widget.reversed && value != null) {
       value = !value;
     }