+++ pref/lib/src/switch.dart
@@ -79,7 +79,14 @@ class _PrefSwitchState extends State<PrefSwitch> {
 
   @override
   Widget build(BuildContext context) {
-    var value = PrefService.of(context).getBool(widget.pref);
+    bool value;
+
+    try {
+      value = PrefService.of(context).get(widget.pref);
+    } catch (e) {
+      print('Unable to load the value: $e');
+    }
+
     if (widget.reversed && value != null) {
       value = !value;
     }