+++ pref/lib/src/checkbox.dart
@@ -65,10 +65,7 @@ class _PrefCheckboxState extends State<PrefCheckbox> {
   }
 
   Future<void> _onChange(bool value) async {
-    setState(() {
-      PrefService.of(context)
-          .setBool(widget.pref, widget.reversed ? !value : value);
-    });
+    PrefService.of(context).set(widget.pref, widget.reversed ? !value : value);
 
     if (widget.onChange != null) {
       widget.onChange(value);