+++ pref/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `reversed` attributes to boolean widgets
 - Improve dropdown layout
+- Prevent crash if the values are not the right type
 
 ## 1.0.0