+++ pref/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `reversed` attributes to boolean widgets
 - Improve dropdown layout
 - Prevent crash if the values are not the right type
+- Remove useless typed methods
 
 ## 1.0.0